requirements.txt
# Core ML libraries
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.2+cu121
transformers>=4.36.0
datasets>=2.14.0
tokenizers>=0.15.0
accelerate>=0.24.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Training utilities
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
pypdf>=4.0.0
python-docx>=1.1.0

# Configuration and utilities
pyyaml>=6.0.1
omegaconf>=2.3.0
hydra-core>=1.3.0

# Evaluation metrics
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.0
